core: move Telemetry, defineLogger, and selectReporters to CLI
Move Telemetry (OTLP tracing), defineLogger, Logger types (Logger,
LoggerContext, LoggerOptions, UserLogger), and selectReporters from
@kubb/core to @kubb/cli. These exports were only ever consumed by the
CLI; programmatic users of createKubb are unaffected.

- Telemetry.ts and Telemetry.test.ts move to packages/cli/src/
- defineLogger.ts moves to packages/cli/src/loggers/defineLogger.ts
  (imports KubbHooks from @kubb/core; logLevel stays in core)
- logLevel const moves from defineLogger.ts into createReporter.ts
- selectReporters moves from createReporter.ts into cli loggers/utils.ts
- OTLP_ENDPOINT moves from core constants into cli constants
- All four CLI Telemetry import sites updated to local paths
- validate/run.test.ts mock updated for the new import path
- reporters/cliReporter.ts and fileReporter.test.ts import logLevel
  from createReporter.ts instead of the removed defineLogger.ts
- Adds changeset for minor bumps to @kubb/core and @kubb/cli

Closes #3530
