+++ packages/cli/src/Telemetry.test.ts
