+++ packages/cli/src/loggers/defineLogger.ts
@@ -1,18 +1,5 @@
 import type { AsyncEventEmitter } from '@internals/utils'
-import type { KubbHooks } from './types.ts'
-
-/**
- * Numeric log-level thresholds used internally to compare verbosity.
- *
- * Higher numbers are more verbose.
- */
-export const logLevel = {
-  silent: Number.NEGATIVE_INFINITY,
-  error: 0,
-  warn: 1,
-  info: 3,
-  verbose: 4,
-} as const
+import type { KubbHooks } from '@kubb/core'
 
 /**
  * Options accepted by a logger's `install` callback.
@@ -22,7 +9,7 @@ export type LoggerOptions = {
    * Output verbosity. Use the `logLevel` constants exported from `@kubb/core`
    * (`silent`, `error`, `warn`, `info`, `verbose`, `debug`).
    */
-  logLevel: (typeof logLevel)[keyof typeof logLevel]
+  logLevel: number
 }
 
 /**
@@ -56,7 +43,7 @@ export type UserLogger<TOptions extends LoggerOptions = LoggerOptions> = Logger<
  *
  * @example
  * ```ts
- * import { defineLogger } from '@kubb/core'
+ * import { defineLogger } from '@kubb/cli'
  *
  * export const myLogger = defineLogger({
  *   name: 'my-logger',
