+++ packages/cli/src/runners/mcp/run.ts
@@ -1,7 +1,7 @@
 import process from 'node:process'
 import { styleText } from 'node:util'
 import { getErrorMessage } from '@internals/utils'
-import { Telemetry } from '@kubb/core'
+import { Telemetry } from '../../Telemetry.ts'
 import type * as McpModule from '@kubb/mcp'
 
 type McpOptions = {
