+++ packages/cli/src/runners/validate/run.test.ts
@@ -5,7 +5,7 @@ vi.mock('@internals/utils', () => ({
   getErrorMessage: vi.fn((error: unknown) => (error instanceof Error ? error.message : String(error))),
 }))
 
-vi.mock('@kubb/core', () => ({
+vi.mock('../../Telemetry.ts', () => ({
   Telemetry: {
     build: vi.fn((payload: object) => payload),
     send: vi.fn(async () => undefined),
