+++ packages/core/src/createReporter.ts
@@ -1,6 +1,18 @@
-import type { logLevel } from './defineLogger.ts'
 import type { Config } from './createKubb.ts'
 import type { Diagnostic } from './diagnostics.ts'
+
+/**
+ * Numeric log-level thresholds used internally to compare verbosity.
+ *
+ * Higher numbers are more verbose.
+ */
+export const logLevel = {
+  silent: Number.NEGATIVE_INFINITY,
+  error: 0,
+  warn: 1,
+  info: 3,
+  verbose: 4,
+} as const
 
 /**
  * A built-in reporter that renders a run's output, independent of the live logger view.
@@ -121,27 +133,3 @@ export function createReporter<T = void>(reporter: UserReporter<T>): Reporter {
   }
 }
 
-/**
- * Picks the reporters whose `name` matches one of `names`, in the order the names are given.
- * The config carries every available reporter, and the host selects which to activate by name
- * (the CLI maps `--reporter` to this). Duplicate names and names without a matching reporter are
- * skipped.
- */
-export function selectReporters(reporters: ReadonlyArray<Reporter>, names: ReadonlyArray<string>): Array<Reporter> {
-  const seen = new Set<string>()
-  const selected: Array<Reporter> = []
-
-  for (const name of names) {
-    if (seen.has(name)) {
-      continue
-    }
-    seen.add(name)
-
-    const reporter = reporters.find((candidate) => candidate.name === name)
-    if (reporter) {
-      selected.push(reporter)
-    }
-  }
-
-  return selected
-}
