+++ packages/core/src/index.ts
@@ -4,15 +4,13 @@ export { createAdapter } from './createAdapter.ts'
 export { createCache } from './createCache.ts'
 export { Diagnostics } from './diagnostics.ts'
 export { createKubb } from './createKubb.ts'
-export { createReporter, selectReporters } from './createReporter.ts'
+export { createReporter, logLevel } from './createReporter.ts'
 export { cliReporter } from './reporters/cliReporter.ts'
 export { fileReporter } from './reporters/fileReporter.ts'
 export { jsonReporter } from './reporters/jsonReporter.ts'
-export { Telemetry } from './Telemetry.ts'
 export { createRenderer } from './createRenderer.ts'
 export { createStorage } from './createStorage.ts'
 export { defineGenerator } from './defineGenerator.ts'
-export { defineLogger, logLevel } from './defineLogger.ts'
 export { defineMiddleware } from './defineMiddleware.ts'
 export { defineParser } from './defineParser.ts'
 export { definePlugin } from './definePlugin.ts'
