+++ packages/core/src/types.ts
@@ -49,7 +49,6 @@ export type { Renderer, RendererFactory } from './createRenderer.ts'
 export type { Storage } from './createStorage.ts'
 export type { FileProcessorHooks, ParsedFile } from './FileProcessor.ts'
 export type { Generator, GeneratorContext } from './defineGenerator.ts'
-export type { Logger, LoggerContext, LoggerOptions, UserLogger } from './defineLogger.ts'
 export type { Middleware } from './defineMiddleware.ts'
 export type { Parser } from './defineParser.ts'
 export type { Exclude, Group, Include, Output, OutputMode, OutputOptions, Override } from './definePlugin.ts'
